% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maaslin3_utils.R
\name{add_taxonomy_to_maaslin3}
\alias{add_taxonomy_to_maaslin3}
\title{Add taxonomy to maaslin3 results}
\usage{
add_taxonomy_to_maaslin3(maaslin3_res, input.tse, taxLevel = "Guess")
}
\arguments{
\item{maaslin3_res}{\code{maaslin3 raw output list} after running the model.}

\item{input.tse}{The \code{TreeSummarizedExperiment} microbiome input data 
used for modeling}

\item{taxLevel}{\code{character} specifying the taxonomic level. Default is 
"Guess", which tries to guess based on the taxa prefixes (eg: Species = s__).}
}
\value{
A \code{list} of \code{data.frame}s, one for the abundance and one 
for the prevalence results, with taxonomy in the first columns and maaslin3 
results following. Taxa not tested are also included for completeness.
}
\description{
This function adds taxonomies to raw maaslin3 results by using the original
input `TreeSummarizedExperiment` microbiome input data. So far, it is designed
to work only with MetaPhlAn4 output tables
}
\examples{

library(maaslin3)
WallenZD_2022.tse <- mia::importMetaPhlAn(
file = system.file("extdata",
                   "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                   package = "biobakeryUtils"),
col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", package = "biobakeryUtils"))

maaslin3_results_raw <- maaslin3(
  input_data = WallenZD_2022.tse, 
  formula = ~ Case_status + Sex + Age_at_collection, 
  transform = "LOG", 
  output = tempdir(),
  normalization = "TSS", # scale values per samples from 0 to 1
  min_prevalence = 0.05, 
  min_abundance = 0.001, 
  standardize = FALSE, 
  verbosity = "ERROR" 
)

maaslin3_with_taxonomy <- add_taxonomy_to_maaslin3(maaslin3_res = maaslin3_results_raw, input.tse = WallenZD_2022.tse, taxLevel = "Guess")

head(maaslin3_with_taxonomy$abundance)
}
