% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPhyloTree_to_mpa_tse.R
\name{AddPhyloTree_to_mpa_tse}
\alias{AddPhyloTree_to_mpa_tse}
\title{This function downloads a cleaned metagenomic phylogenetic tree from the 
CHOCOPhlAn database and incorporates it as the `rowTree` in a 
`TreeSummarizedExperiment` object that currently lacks one. It supports 
adding a special "UNCLASSIFIED" tip to the root of the tree with branch 
length 1 if such reads are present in the data.}
\usage{
AddPhyloTree_to_mpa_tse(data.tse, CHOCOPhlAn_version = "latest")
}
\arguments{
\item{data.tse}{A \code{TreeSummarizedExperiment} object without an existing
`rowTree`.}

\item{CHOCOPhlAn_version}{A \code{character} string specifying the version 
of the CHOCOPhlAn tree to download.
Supported versions include "latest" (default), "202503", "202403", and 
"202307". These correspond to dated releases available at:
\url{http://cmprod1.cibio.unitn.it/biobakery4/metaphlan_databases/}.}
}
\value{
A \code{TreeSummarizedExperiment} object identical to the input but 
with the `rowTree` slot populated with the downloaded and pruned phylogenetic
tree matching the taxa in `data.tse`.
}
\description{
This function downloads a cleaned metagenomic phylogenetic tree from the 
CHOCOPhlAn database and incorporates it as the `rowTree` in a 
`TreeSummarizedExperiment` object that currently lacks one. It supports 
adding a special "UNCLASSIFIED" tip to the root of the tree with branch 
length 1 if such reads are present in the data.
}
\details{
The function performs the following steps:
\itemize{
  \item Checks if `rowTree` is already present; if so, returns the input 
  unchanged.
  \item Downloads the specified CHOCOPhlAn tree version.
  \item Renames tree tips to the format "t__SGB<tip.label>" as per Biobakery
  community guidelines.
  \item Adds a tip labeled "UNCLASSIFIED" at the root if the dataset contains
  such reads.
  \item Prunes the tree to keep only tips present in the dataset.
  \item Reorders the `TreeSummarizedExperiment` rows to match the tree tip
  order.
  \item Assigns the pruned tree to the `rowTree` slot.
}
}
\examples{
\dontrun{
WallenZD_2022.tse <- mia::importMetaPhlAn(
  file = system.file("extdata", "WallenZD_2022_metaphlan3_profiles.tsv.bz2", package = "biobakeryUtils"),
  col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", package = "biobakeryUtils")
)
# Assuming tse is a TreeSummarizedExperiment without a rowTree
tse_with_tree <- AddPhyloTree_to_mpa_tse(WallenZD_2022.tse, CHOCOPhlAn_version = "201903")
}
}
