% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_taxonomy_levels}
\alias{all_taxonomy_levels}
\title{A vector of taxonomies and their typical prefixes in metaphlan}
\format{
A character vector with 8 elements:
\describe{
  \item{all_taxonomy_levels}{The taxonomy level with capitalized 1st letter}
}
}
\usage{
all_taxonomy_levels
}
\description{
This vector contains taxonomies from Kingdom to SGB level. the vector is
generated as : all_taxonomy_levels <- c( k__ = "Kingdom", p__ = "Phylum",
  c__ = "Class", o__ = "Order", f__ = "Family", g__ = "Genus", 
 s__ = "Species", t__ = "SGB")
}
\keyword{datasets}
