% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bedarf_pMD_raw_HUMAnN_pwy.tse}
\alias{Bedarf_pMD_raw_HUMAnN_pwy.tse}
\title{Bedarf et al. HUMAnN pathway abundance data (unstratified)}
\format{
A \code{TreeSummarizedExperiment} object.
}
\source{
\code{parkinsonsMetagenomicData} Bioconductor package.
}
\usage{
data(Bedarf_pMD_raw_HUMAnN_pwy.tse)
}
\description{
A TreeSummarizedExperiment containing HUMAnN pathabundance_unstratified
profiles from the Bedarf study, retrieved via parkinsonsMetagenomicData.
}
\keyword{datasets}
