% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bedarf_pMD_raw_MetaPhlAn.tse}
\alias{Bedarf_pMD_raw_MetaPhlAn.tse}
\title{Bedarf et al. MetaPhlAn relative abundance data}
\format{
A \code{TreeSummarizedExperiment} object.
}
\source{
\code{parkinsonsMetagenomicData} Bioconductor package.
}
\usage{
data(Bedarf_pMD_raw_MetaPhlAn.tse)
}
\description{
A TreeSummarizedExperiment containing MetaPhlAn relative abundance
profiles from the Bedarf study, retrieved via parkinsonsMetagenomicData.
}
\keyword{datasets}
