% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{colDataSpecs}
\alias{colDataSpecs}
\title{Specify classes of column data
Most importantly, the levels of a factor}
\usage{
colDataSpecs(tse)
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} Object}
}
\value{
A \code{data.frame}
}
\description{
Specify classes of column data
Most importantly, the levels of a factor
}
\examples{
 library(TreeSummarizedExperiment)
 data("tinyTree")

 # the count.mat table
 count.mat <- matrix(rpois(100, 50), nrow = 10)

 rownames(count.mat) <- c(tinyTree$tip.label)

 colnames(count.mat) <- paste("C_", 1:10, sep = "_")

 # The sample information
sampC <- data.frame(
  condition = rep(c("control", "trt"), each = 5),
  gender = sample(x = 1:2, size = 10, replace = TRUE)
  )

rownames(sampC) <- colnames(count.mat)

 # build a TreeSummarizedExperiment object
 tse <- TreeSummarizedExperiment(
   assays = list("counts" = count.mat),
   colData = sampC,
   rowTree = tinyTree
   )
   
 tse
}
