% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{collapse_taxonomy}
\alias{collapse_taxonomy}
\title{Taxonomy from data.frame to vector}
\usage{
collapse_taxonomy(taxonomy.df, sep = "|")
}
\arguments{
\item{taxonomy.df}{\code{data.frame} with taxonomic levels on columns and 
individual taxa on rows. Taxa levels must be ordered from Kingdom/Domain to 
make sense, but the function does not check that}

\item{sep}{\code{character} specifying the separator avoid 
`c("_", "-", " ", ",", "\\t")` separators, because they will likely be 
present in the taxonomy table already or because it would conflict with 
saving the taxonomy table in .csv or .tsv format}
}
\value{
A \code{character} of complete taxonomies separated by `sep`
}
\description{
Taxonomy from data.frame to vector
}
\examples{
example_taxonomies_Species.chr <- c(
"k__Bacteria|p__Firmicutes|c__CFGB8118|o__OFGB8118|f__FGB8118|g__GGB29005|s__GGB29005_SGB41723",
"k__Bacteria|p__Firmicutes|c__Clostridia|o__Eubacteriales|f__Lachnospiraceae|g__Lachnospiraceae_unclassified|s__Lachnospiraceae_unclassified_SGB41492",
"k__Bacteria|p__Bacteria_unclassified|c__Bacteria_unclassified|o__Bacteria_unclassified|f__Bacteria_unclassified|g__GGB28904|s__GGB28904_SGB41595"
)

expanded_taxonomies.df <- expand_taxonomy(example_taxonomies_Species.chr)

recollapsed_taxonomies.chr <- collapse_taxonomy(expanded_taxonomies.df)

identical(recollapsed_taxonomies.chr, example_taxonomies_Species.chr)
}
