% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R, R/wrangle_metaphlan.R
\name{complete_unknown_taxonomy}
\alias{complete_unknown_taxonomy}
\title{Complete Unknown taxonomy
 
This is a utility function for the `metaphlan_wrangle function`, which needs
it in case it finds taxa that are unique at higher taxonomies levels than the 
one requested (e.g. a Genus without any species mapped to it. maybe it is still 
a valuable genus, so it has to be repeated with the same name at the taxonomic level below).
This is also valuable because another approach would discard it, and the total
relative abundances estimated would not add up.}
\usage{
complete_unknown_taxonomy(x, tax_lvl_int = 8)

complete_unknown_taxonomy(x, tax_lvl_int = 8)
}
\arguments{
\item{x}{\code{character},  the taxonomy as it appears in the first column of a metaphlan table}

\item{tax_lvl_int}{\code{character} the level at which the user wants the taxonomies aggregated. The default goes all the way to the SGB level (`8`). Species level is `7`, Genus is `6` and so on.}
}
\value{
\code{character}, the same as the input vector, with the lowest 
taxonomy repeated down to the `tax_lvl_int` chosen

\code{character}, the same as the input vector, with the lowest taxonomy repeated down to the `tax_lvl_int` chosen
}
\description{
Complete Unknown taxonomy
 
This is a utility function for the `metaphlan_wrangle function`, which needs
it in case it finds taxa that are unique at higher taxonomies levels than the 
one requested (e.g. a Genus without any species mapped to it. maybe it is still 
a valuable genus, so it has to be repeated with the same name at the taxonomic level below).
This is also valuable because another approach would discard it, and the total
relative abundances estimated would not add up.

Complete Unknown taxonomy
 
This is a utility function for the `metaphlan_wrangle function`, which needs
it in case it finds taxa that are unique at higher taxonomies levels than the 
one requested (e.g. a Genus without any species mapped to it. maybe it is still 
a valuable genus, so it has to be repeated with the same name at the taxonomic level below).
This is also valuable because another approach would discard it, and the total
relative abundances estimated would not add up.
}
\examples{

example_taxonomies <- c(
"k__Bacteria|p__Firmicutes|c__CFGB1347", 
"k__Bacteria|p__Actinobacteria", 
"k__Bacteria|p__Firmicutes|c__Clostridia|o__Eubacteriales|f__Eubacteriales_Family_XIII_Incertae_Sedis|g__Lentihominibacter|s__Lentihominibacter_faecis")

# complete taxonomies to species level (level of taxonomic depth)
complete_unknown_taxonomy(example_taxonomies, tax_lvl_int = 7)


example_taxonomies <- c("k__Bacteria|p__Firmicutes|c__CFGB1347", "k__Bacteria|p__Actinobacteria", "k__Bacteria|p__Firmicutes|c__Clostridia|o__Eubacteriales|f__Eubacteriales_Family_XIII_Incertae_Sedis|g__Lentihominibacter|s__Lentihominibacter_faecis")

# complete taxonomies to species level (level of taxonomic depth)
complete_unknown_taxonomy(example_taxonomies, tax_lvl_int = 7)


}
