% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{expand_taxonomy}
\alias{expand_taxonomy}
\title{Taxonomy from vector to data.frame}
\usage{
expand_taxonomy(taxonomy.chr, sep = "|", row.names = "short")
}
\arguments{
\item{taxonomy.chr}{A \code{character} vector with the taxonomy. They should 
come from a pre-cleaned taxonomy table, such as after `mia::importMetaphlan`}

\item{sep}{A \code{character} specifying the separator to use to split 
`taxonomy.chr`}

\item{row.names}{Either `'short'`, `'long'`, or `'none'`. Default is 
`'short'`, which returns the highest taxonomic resolution the input 
`taxonomy.chr` has for that row. `'long'` returns `taxonomy.chr` as column 
names. `'none'` returns a data.frame without rown ames}
}
\value{
A \code{data.frame} of taxonomies
}
\description{
Expand a character vector of taxonomies into a full taxonomy table based on 
the separator provided and the prefixes at each level. Tested on metaphlan4
output tables
}
\examples{
example_taxonomies_SGB <- c(
  "k__Bacteria|p__Firmicutes|c__CFGB8118|o__OFGB8118|f__FGB8118|g__GGB29005|s__GGB29005_SGB41723|t__SGB41723",
  "k__Bacteria|p__Firmicutes|c__Clostridia|o__Eubacteriales|f__Lachnospiraceae|g__Lachnospiraceae_unclassified|s__Lachnospiraceae_unclassified_SGB41492|t__SGB41492",
  "k__Bacteria|p__Bacteria_unclassified|c__Bacteria_unclassified|o__Bacteria_unclassified|f__Bacteria_unclassified|g__GGB28904|s__GGB28904_SGB41595|t__SGB41595"
)

expand_taxonomy(example_taxonomies_SGB)

example_taxonomies_Species <- c(
  "k__Bacteria|p__Firmicutes|c__CFGB8118|o__OFGB8118|f__FGB8118|g__GGB29005|s__GGB29005_SGB41723",
  "k__Bacteria|p__Firmicutes|c__Clostridia|o__Eubacteriales|f__Lachnospiraceae|g__Lachnospiraceae_unclassified|s__Lachnospiraceae_unclassified_SGB41492",
  "k__Bacteria|p__Bacteria_unclassified|c__Bacteria_unclassified|o__Bacteria_unclassified|f__Bacteria_unclassified|g__GGB28904|s__GGB28904_SGB41595"
)

expand_taxonomy(example_taxonomies_Species)
}
