% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{get_alternate_SGB_names}
\alias{get_alternate_SGB_names}
\title{Find SGB alternate names}
\usage{
get_alternate_SGB_names(SGB_id, version = "202403")
}
\arguments{
\item{SGB_id}{\code{character} of lenght 1 with SGB code}

\item{version}{\code{character} of date in the format %Y%m that specifies
the CHOCOPhlAn database version}
}
\value{
A \code{character} vector with the alternate names in CHOCOPhlAn
}
\description{
Find SGB alternate names
}
\examples{
SGB_with_no_altNames <- get_alternate_SGB_names("t__SGB51429", version = "202403")
SGB_with_no_altNames

SGB_with_altNames <- get_alternate_SGB_names("t__SGB13547", version = "202403")
SGB_with_altNames
}
