% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{get_mpa_full_taxonomy}
\alias{get_mpa_full_taxonomy}
\title{Upload full taxonomy data from MetaPhlAn database}
\usage{
get_mpa_full_taxonomy(version = "202403")
}
\arguments{
\item{version}{\code{character} with CHOCOPhlAn database version as found in 
unitn cmprod1}
}
\value{
\code{data.frame, data.table} object, with columns being `SGB`, 
`Number of alterantive columns`, `list of alternative names`
}
\description{
Upload full taxonomy data from MetaPhlAn database
}
\examples{

head(get_mpa_full_taxonomy("202403"))
}
