% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reads_processed.R
\name{get_reads_processed}
\alias{get_reads_processed}
\title{Get reads processed by MetaPhlAn 3 and 4}
\usage{
get_reads_processed(metaphlan_profile_paths)
}
\arguments{
\item{metaphlan_profile_paths}{\code{character}, The path to the metaphlan 
profile.}
}
\value{
A \code{integer} vector containing the number of reads processed
}
\description{
A simple function to read in the .gz files cached with 
`parkinsonsMetagenomicData::cacheMetagenomicData()`
}
\examples{

\dontrun{
# this doesn't work yet, at some point I will put an example file into 
# extdata
  path_to_mpa <- system.file("mpa_single_file.tsv.gz", "extdata", 
  package = "biobakeryUtils")
  
  get_reads_processed(path_to_mpa)
}
}
