% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureStats.R
\name{getFeatureStats}
\alias{getFeatureStats}
\title{get Features stats}
\usage{
getFeatureStats(tse, assay.type = NULL)
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} to calculate the stats on}

\item{assay.type}{\code{character} of the assay to calculate the estimates on (default: `NULL`)}
}
\value{
\code{data.frame} with one row per feature and some common useful 
summary stats in columns
}
\description{
get Features stats
}
\examples{
library(mia)
WallenZD_2022.tse <- mia::importMetaPhlAn(
file = system.file("extdata",
                   "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                   package = "biobakeryUtils"),
col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", 
package = "biobakeryUtils"))

Wallen_stats.df <- getFeatureStats(WallenZD_2022.tse)

}
