% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetagenomicData.R
\name{getMetaPhlAn_run_info}
\alias{getMetaPhlAn_run_info}
\title{Extract run info of a MetaPhlAn run}
\usage{
getMetaPhlAn_run_info(files_to_read)
}
\arguments{
\item{files_to_read}{A \code{character} vector with the metaphlan profile already 
read into R with readLines}
}
\value{
A \code{tibble} of per-sample MetaPhlAn run information (db version, code, reads)
}
\description{
Extract run info of a MetaPhlAn run
}
\examples{

\dontrun{
mpaProfile <- readLines("path/to/metaphlan_run.tsv")

mpa_run_info <- getMetaPhlAn_run_info(mpaProfile)

}

}
