% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_utils.R
\name{limmaFit_TSE}
\alias{limmaFit_TSE}
\title{limma wrapper for biobakery data in TreeSummarizedExperiment format}
\usage{
limmaFit_TSE(tse, assay.type, formula, coef, p.adj_method = "BH")
}
\arguments{
\item{tse}{\code{(Tree)SummarizedExperiment}. Input data with available rowData}

\item{assay.type}{\code{character}. The assay to test with limma}

\item{formula}{\code{formula}. Model to pass to `model.matrix` before lmFit}

\item{coef}{\code{character}. the coefficient as seen in one of `limma$coef` columns names.}

\item{p.adj_method}{\code{character} One of the methods allowed by `stats::p.adjust`}
}
\value{
a \code{data.frame} with a lot of feature information along with 
limma's summary statistics
}
\description{
NB: This particular wrapper does not perforom limma::voom, but simply the 
`lmFit` function with a given formula. It does not perform transformations
either. IMPORTANT: taxa stats are calculated on the first assay, which is 
expected to be a `abundance/relabundance` as clean as possible from either
MetaPhlAn or HUMAnN.
}
\examples{
library(mia)
data(enterotype)
enterotype <- transformAssay(enterotype, 
  assay.type = "counts", 
  method = "clr", 
  pseudocount = TRUE)

no_NA_variables <- complete.cases(colData(enterotype)[,c("Enterotype","Age")])
enterotype_noNAs <- enterotype[,no_NA_variables]
limmaResults <- limmaFit_TSE(enterotype_noNAs, 
  assay.type = "clr", 
  formula = ~ Enterotype + Age, 
  coef = "Enterotype3")

str(limmaResults)
}
