% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetagenomicData.R
\name{loadMetagenomicData}
\alias{loadMetagenomicData}
\title{Rewriting of loadMetagenomicData in parkinsonsMetagenomicData}
\usage{
loadMetagenomicData(cache_table)
}
\arguments{
\item{cache_table}{\code{data.frame} containing local paths where data have been 
cached with `cacheMetagenomicData`}
}
\value{
A \code{TreeSummarizedExperiment} (without phylogenetic tree), as 
generated by `mia:importMetaPhlAn`.
}
\description{
This function should alegedly work faster and be more flexible toward
empty files. Additionally, it makes use of mia::importMetaphlan, which 
is constantly maintained tested. **IMPORTANT**: to work, the function must
have the right to write in its temporary directory. This code may not work in
some cloud computing environments.
}
\examples{

# careful, this example takes a few minutes to run
\dontrun{
library(parkinsonsMetagenomicData)

human_parkinson_data <- dplyr::filter(sampleMetadata, !grepl("Mazmanian", study_name))

cached_profiles <- suppressMessages(cacheMetagenomicData(human_parkinson_data$uuid, data_type = "relative_abundance"))

is_downloaded <- sapply(cached_profiles$cache_path, file.exists)

table(is_downloaded)

cache_table <- cached_profiles[is_downloaded,]

# this one fails with this subset of samples
data.se <- loadMetagenomicData(cache_table)

# this one does not fail
data.tse <- loadMetagenomicData(cache_table)
}
}
