% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalysisUtils.R
\name{metaAnalyze}
\alias{metaAnalyze}
\title{Meta-Analyze multi-group}
\usage{
metaAnalyze(
  res.df,
  grouping.col = "study_name",
  beta.col = "Beta",
  se.col = "SE",
  method = "REML",
  test = "knha",
  minStudies.n = 6,
  p.adj_method = "BH"
)
}
\arguments{
\item{res.df}{\code{data.frame} with long-format meta-analysis results}

\item{grouping.col}{\code{character} variable that was used to group data
which were later meta-analyzed}

\item{beta.col}{\code{character}. Column name of the `beta` estimates}

\item{se.col}{\code{character}. Column name of the `SE` estimates}

\item{method}{\code{character}. Meta-analysis method as found in the omonimous
parameter in metafor::rma}

\item{test}{\code{character}. Meta-analysis test as found in the omonimous
parameter in metafor::rma}

\item{minStudies.n}{\code{integer}. Minimum number of studies to retain a 
feature for meta-analysis}

\item{p.adj_method}{\code{character}. One of the possible options in 
`p.adjust`}
}
\value{
A \code{list} of 3 elements: (1) "Beta_matrix"; (2) "SE_matrix"; 
 (3) "MetaAnalysis". The first 2 are the matrices used to calculate the meta-
 analyses, while the 3rd one is a `data.frame` of all results together
}
\description{
Meta-Analyze multi-group
}
\examples{
library(curatedMetagenomicData)
library(mia)
library(magrittr)
library(dplyr)

sex_meta_analysis <- sampleMetadata \%>\%
  # keep cases where we know sex and that are not longitudinal samples
  filter(
    !is.na(sampleMetadata$gender),
    is.na(sampleMetadata$visit_number),
    body_site == "stool", 
    disease == "healthy"
  )

# restrict to datasets with at least 30 males and 30 females
ok_dataset <- sex_meta_analysis \%>\%
  group_by(study_name) \%>\%
  reframe(
    nMale = sum(gender == "male"),
    nFemale = sum(gender == "female")
  ) \%>\%
  filter(nMale > 30, nFemale > 30)

sex_meta_analysis <- filter(sex_meta_analysis, study_name \%in\% ok_dataset$study_name)
sex_meta_analysis.se <- returnSamples(sex_meta_analysis, 
                                      dataType = "relative_abundance", counts = TRUE)

# make a copy of the assay, it actually has counts, as requested
# testable with `range(assay(sex_meta_analysis.se))`
assay(sex_meta_analysis.se, "counts") <- assay(sex_meta_analysis.se, 
                                               "relative_abundance")


# filter by prevalence, at least 5 counts in each cell, at least 1\% of 
# the samples
sex_meta_analysis.se <- mia::subsetByPrevalent(sex_meta_analysis.se, 
  assay.type = "counts", detection = 5, prevalence = 0.01)
# fix error of same feature in multiple lines by aggregating

sex_meta_analysis.se <- mia::agglomerateByRank(sex_meta_analysis.se, 
  "species")


singledataset_analysis_sex.df <- sex_meta_analysis.se \%>\%
  splitOn(group = "study_name") \%>\%
  lapply(function(x.tse) {
    x.tse <- x.tse[rowSums(assay(x.tse, "relative_abundance")) > 0,]
    return(limmaTSE(tse = x.tse, assay.type = "counts", 
                      voom.transform = TRUE, 
                      formula = ~ gender, coef = "gendermale", 
                      p.adj_method = "BH"))
  }
  ) \%>\% bind_rows(.id = "study_name")

meta_analysis_results <- metaAnalyze(res.df = singledataset_analysis_sex.df,
                                     grouping.col = "study_name", 
                                     beta.col = "Beta", 
                                     se.col = "SE", method = "REML", 
                                     test = "knha", 
                                     minStudies.n = 6, 
                                     p.adj_method = "BH")

str(meta_analysis_results)
}
