% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalysisUtils.R
\name{metaAnalyzeSingle}
\alias{metaAnalyzeSingle}
\title{Single-outcome Meta-analysis with metafor}
\usage{
metaAnalyzeSingle(yi, sei, method = "REML", test = "knha")
}
\arguments{
\item{yi}{\code{numeric} vector of effect sizes `beta`}

\item{sei}{\code{numeric} vector of standard errors of `beta`}

\item{method}{\code{chracter}. One of the options from the `method` parameter
in metafor::rma. (Default: 'REML')}

\item{test}{\code{chracter}. One of the options from the `test` parameter
in metafor::rma. (Default: 'knha', which is suggested for meta-analyses 
with up to 20 datasets.}
}
\value{
\code{numeric} vector containing the meta-analysis basic stats:
StudiesTested: number of studies with feature, Beta.pool: random effects beta
SE.pool: standard error of the Beta.pool point esimate; CI.L: 95% lower bound 
confidence interval of the Beta and SE; CI.R: upper 95% bound confidence 
interval; P.Value: meta-analysis p-value estimated with the method in `test`;
Het.tau2: tau-squared heterogeneity estimate, Het.I2: I-squared heterogeneity
estimate Het.Pval: P-value associated with heterogeneity.
}
\description{
This function is a simple wrapper for what I need of a typical meta-analysis
workflow for micobiome data. This function is called by \code{metaAnalyze},
which generalizes its use.
}
\examples{
library(metafor)
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)
metaAnalyzeSingle(yi = dat$yi, sei = dat$vi, method = "REML")
}
