% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERMANOVA_to_table.R
\name{PERMANOVA_to_table}
\alias{PERMANOVA_to_table}
\title{PERMANOVA to table}
\usage{
PERMANOVA_to_table(tse, name = "permanova")
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} object that has been already populated with `mia::addPERMANOVA`}

\item{name}{\code{character} specifying the 'name' paramenter that was passed to `mia::addPERMANOVA`}
}
\value{
A \code{data.frame} with permanova and betadisper + permutest statistics
ready to be reported or saved as tables.
}
\description{
A utility function that produces a data.frame with PERMANOVA and betadisper
statistics from a TreeSummarizedExperiment object
}
\examples{
library(mia)
data(GlobalPatterns)
tse <- GlobalPatterns

# Apply relative transformation
tse <- transformAssay(tse, method = "relabundance")
# Perform PERMANOVA
tse <- addPERMANOVA(
  tse,
  assay.type = "relabundance",
  method = "bray",  
  formula = x ~ SampleType,
  permutations = 99
)
# The results are stored to metadata
metadata(tse)[["permanova"]]

# PERMANOVA_to_table makes it faster to retrieve all pieces of it in the
# tse object
permanova.df <- PERMANOVA_to_table(tse, name = "permanova")
permanova.df
}
