% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{pMD_enhance}
\alias{pMD_enhance}
\title{Enhance parkinsonsMetagenomicData output}
\usage{
pMD_enhance(input.tse, sampleMetadata.df, data_type)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of
\code{loadMetagenomcData} or \code{returnSamples}}

\item{sampleMetadata.df}{\code{data.frame}. A data.frame to swap the default 
with. NB: it must have a `uuid` column with valid uuids
\code{loadMetagenomcData} or \code{returnSamples}}

\item{data_type}{\code{character}. that got retrieved using 
\code{loadMetagenomcData} or \code{returnSamples}
to the \code{TreeSummarizedExperiment}? Works only with MetaPhlAn input 
(`data_type = "relative_abundance"`)}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Formats and adds pieces to select outputs of parkinsonsMetagenomicData
}
\examples{

library(parkinsonsMetagenomicData)
data("sampleMetadata", package = "parkinsonsMetagenomicData")

data_types <- c("relative_abundance", 
"pathabundance_unstratified", "genefamilies_unstratified")

example_metadata.df <- sampleMetadata[grepl("Bedarf", 
sampleMetadata$study_name),]

# MetaPhlAn

tse_basic <- returnSamples(example_metadata.df, data_type = data_types[1])

tse_enhanced <- pMD_enhance(tse_basic, 
  sampleMetadata.df = example_metadata.df, 
  data_type = data_types[1])

# HUMAnN pathways
tse_basic <- returnSamples(example_metadata.df, data_type = data_types[2])

tse_enhanced <- pMD_enhance(tse_basic, 
  sampleMetadata.df = example_metadata.df, 
  data_type = data_types[2])

tse_basic <- returnSamples(example_metadata.df, data_type = data_types[3])

tse_enhanced <- pMD_enhance(tse_basic, 
  sampleMetadata.df = example_metadata.df,
  data_type = data_types[3])


}
