% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_enhance.R
\name{pMD_enhance_HUMAnN_pwy}
\alias{pMD_enhance_HUMAnN_pwy}
\title{Enhance HUMAnN unstratified pathways data}
\usage{
pMD_enhance_HUMAnN_pwy(input.tse)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData} or \code{returnSamples}. works exclusively for the
\code{data_type = "pathabundance_unstratified"} case}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance HUMAnN unstratified pathways data
}
\examples{
suppressMessages(library(TreeSummarizedExperiment))
data_types <- c("relative_abundance",
"pathabundance_unstratified", "genefamilies_unstratified")

data(Bedarf_pMD_raw_HUMAnN_pwy.tse)

tse_enhanced <- pMD_enhance_HUMAnN_pwy(Bedarf_pMD_raw_HUMAnN_pwy.tse)

tse_enhanced
}
