% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_enhance.R
\name{pMD_enhance_MetaPhlAn}
\alias{pMD_enhance_MetaPhlAn}
\title{Enhance MetaPhlAn data}
\usage{
pMD_enhance_MetaPhlAn(input.tse)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData} or \code{returnSamples}. works exclusively for the
\code{data_type = "relative_abundance"} case}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance MetaPhlAn data
}
\examples{
suppressMessages(library(TreeSummarizedExperiment))
data_types <- c("relative_abundance",
"pathabundance_unstratified", "genefamilies_unstratified")

data(Bedarf_pMD_raw_MetaPhlAn.tse)

tse_enhanced <- pMD_enhance_MetaPhlAn(Bedarf_pMD_raw_MetaPhlAn.tse)

tse_enhanced
}
