% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{rename_rownames.tse}
\alias{rename_rownames.tse}
\title{Enhanced row renaming in a TreeSummarizedExperiment}
\usage{
rename_rownames.tse(data.tse, new.rownames)
}
\arguments{
\item{data.tse}{\code{(Tree)SummarizedExperiment} with or without a 
phylogenetic tree}

\item{new.rownames}{= A \code{character} vector of new rownames. 
IMPORTANT! Make sure the order of old and new row names makes sense before 
running this function}
}
\value{
The same \code{TreeSummarizedExperiment}, but with reordered 
rownames of both `assays` and `rowData`
}
\description{
This includes renaming not only rows of a TSE, but renames tree tips as well
}
\examples{
library(mia)
WallenZD_2022.tse <- mia::importMetaPhlAn(
file = system.file("extdata",
                   "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                   package = "biobakeryUtils"),
col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", 
package = "biobakeryUtils"))

WallenZD_2022_fullLengthNames.tse <- rename_rownames.tse(WallenZD_2022.tse, 
new.rownames = collapse_taxonomy(as.data.frame(rowData(WallenZD_2022.tse))))

head(rownames(WallenZD_2022_fullLengthNames.tse))
head(rownames(WallenZD_2022.tse))
}
