% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_utils.R
\name{reorder_taxa_with_phyloTree_labels}
\alias{reorder_taxa_with_phyloTree_labels}
\title{Reorder rows and tip labels of a TreeSummarizedExperiment}
\usage{
reorder_taxa_with_phyloTree_labels(data.tse)
}
\arguments{
\item{data.tse}{\code{TreeSummarizedExperiment} object. It must contain a 
phylogenetic tree}
}
\value{
\code{TreeSummarizedExperiment} object with identical rownames and 
tip labels order
}
\description{
Reorder rows and tip labels of a TreeSummarizedExperiment
}
\examples{
\dontrun{
# load a tree summarized experiment
 WallenZD_2022.tse <- mia::importMetaPhlAn(
  file = system.file("extdata",
                    "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                    package = "biobakeryUtils"),
  col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", 
  package = "biobakeryUtils")
  )
 # Add phylogenetic tree
 
 # shuffle rownames of the 
 
 # check if they are identical to tree tips labels
 
 tse_reordered <- reorder_taxa_with_phyloTree_labels(tse)
}
}
