% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{replace_colData}
\alias{replace_colData}
\title{Replace colData of a (Tree)SummarizedExperiment}
\usage{
replace_colData(input.se, new.df)
}
\arguments{
\item{input.se}{\code{(Tree)SummarizedExperiment}}

\item{new.df}{\code{data.frame or DataFrame} with rownames corresponding to 
rownames in `rownames(colData(input.se))`. The correct order is coerced.}
}
\value{
\code{(Tree)SummarizedExperiment} with replaced colData
}
\description{
A pipe-friendly wrapper to replace colData from a (Tree)SummarizedExperiment.
By design, if a new colData with more samples is provided, only those that map
to input.se are inserted in the colData slot.
}
\examples{
\dontrun{
library(parkinsonsMetagenomicData)

data_types <- c("relative_abundance", "pathabundance_unstratified", "genefamilies_unstratified")

hf_con <- accessParquetData(data_types = data_types)

example_metadata.df <- dplyr::filter(sampleMetadata, grepl("Bedarf", study_name))

example_metadata.df$ABC_Var <- factor(sample(1:0, nrow(example_metadata.df), replace = TRUE))
rownames(example_metadata.df) <- example_metadata.df$uuid

# HUMAnN gene families
tse_basic <- loadParquetData(
  con = hf_con, 
  data_type = "genefamilies_unstratified", 
  filter_values = list(uuid = example_metadata.df$uuid)
)

"ABC_Var" \%in\% colnames(colData(tse_basic))

tse_newColData <- tse_basic \%>\%
  replace_colData(  new.df = example_metadata.df)

"ABC_Var" \%in\% colnames(colData(tse_newColData))
  
}
}
