% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nearest_taxa.R
\name{search_nearest_taxa}
\alias{search_nearest_taxa}
\title{Find phylogenetically closest taxa to a target}
\usage{
search_nearest_taxa(phyloTree, target_SGB, taxonomy_sep = "|")
}
\arguments{
\item{phyloTree}{\code{phylo} tree that can be handled by `ape` and 
`tidytree` CRAN packages. IMPORTANT: So far, this function was thought to
 work on full taxonomy as tip labels. So before extracting the phyloTree,
 please run `rename_rownames.tse(your_data.tse, collapse_taxonomy(as.data.frame(rowData(your_data.tse))))`}

\item{target_SGB}{\code{character} with SGB name either as "t__SGB12345" or 
as full taxonomy.}

\item{taxonomy_sep}{\code{character} specifying the separation character used
in the tree's tip labels. Default is "|"}
}
\value{
\code{data.frame} with phylogenetically closest taxa based on the 
input phylogenetic tree.
}
\description{
Find phylogenetically closest taxa to a target
}
\examples{
\dontrun{
WallenZD_2022.tse <- mia::importMetaPhlAn(
 file = system.file("extdata",
                   "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                   package = "biobakeryUtils"),
 col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", 
 package = "biobakeryUtils")
 )

# does not work because WallenZD reported data at species level
WallenZD_2022.tse <- AddPhyloTree_to_mpa_tse(
  data.tse = WallenZD_2022.tse, 
  CHOCOPhlAn_version = "202403"
  )
}
}
