% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mediation.R
\name{tidy_mediation}
\alias{tidy_mediation}
\title{Extract summary info from `mediation::mediate` model}
\usage{
tidy_mediation(mediation.model)
}
\arguments{
\item{mediation.model}{\code{mediation} object}
}
\value{
\code{data.frame} with stats as come out of `summary(mediation.model)`
}
\description{
Extract summary info from `mediation::mediate` model
}
\examples{
\dontrun{

#---------- taken from the mediation::mediate first example
library(mediation)
b <- lm(job_seek ~ treat + econ_hard + sex + age, data=jobs)
c <- lm(depress2 ~ treat + job_seek + econ_hard + sex + age, data=jobs)
# Estimation via quasi-Bayesian approximation
contcont <- mediate(b, c, sims=50, treat="treat", mediator="job_seek")
summary(contcont)

# summarized as data.frame
tidy_mediation(contcont)
}
}
