% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maaslin3_utils.R
\name{write_maaslin3_curated_tables}
\alias{write_maaslin3_curated_tables}
\title{Write curated maaslin3 output}
\usage{
write_maaslin3_curated_tables(maaslin3_res_list, out.dir)
}
\arguments{
\item{maaslin3_res_list}{\code{list} of results as come out of 
`add_taxonomy_to_maaslin3`}

\item{out.dir}{\code{character} specifying where to save results}
}
\value{
NULL, nothing is returned. Files are save as 
`all_results_w_taxonomy_abundance.tsv` and 
`all_results_w_taxonomy_prevalence.tsv`
}
\description{
This is a quick function to save data that result from the previous function
called `add_taxonomy_to_maaslin3`
}
\examples{

library(maaslin3)
WallenZD_2022.tse <- mia::importMetaPhlAn(
file = system.file("extdata",
                   "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                   package = "biobakeryUtils"),
col.data = system.file("extdata", "WallenZD_2022_subjMetadata.tsv.bz2", package = "biobakeryUtils"))

maaslin3_results_raw <- maaslin3(
  input_data = WallenZD_2022.tse, 
  formula = ~ Case_status + Sex + Age_at_collection, 
  transform = "LOG", 
  output = tempdir(),
  normalization = "TSS", # scale values per samples from 0 to 1
  min_prevalence = 0.05, 
  min_abundance = 0.001, 
  standardize = FALSE, 
  verbosity = "ERROR" 
)

maaslin3_with_taxonomy <- add_taxonomy_to_maaslin3(maaslin3_res = maaslin3_results_raw, input.tse = WallenZD_2022.tse, taxLevel = "Guess")

head(maaslin3_with_taxonomy$abundance)
head(maaslin3_with_taxonomy$prevalence)

write_maaslin3_curated_tables(maaslin3_with_taxonomy, tempdir())

list.files(tempdir(), pattern = "all_results_w_taxonomy*")
}
