% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{colDataSpecs}
\alias{colDataSpecs}
\title{Specify classes of column data
Most importantly, the levels of a factor}
\usage{
colDataSpecs(tse)
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} Object}
}
\value{
A \code{data.frame}
}
\description{
Specify classes of column data
Most importantly, the levels of a factor
}
\examples{
# Example data ----
library(TreeSummarizedExperiment)
data("tinyTree")

# the count.mat table
set.seed(1234)
count.mat <- matrix(rpois(100, 50), nrow = 10)
rownames(count.mat) <- tinyTree$tip.label
colnames(count.mat) <- paste0("C_", 1:10)

# the sample information
sampC <- data.frame(
  condition = factor(rep(c("control", "trt"), each = 5), 
                     levels = c("trt", "control")),
  gender = factor(sample(c("male", "female"), size = 10, replace = TRUE),
                  levels = c("male", "female"))
)
rownames(sampC) <- colnames(count.mat)

# build a TreeSummarizedExperiment
tse <- TreeSummarizedExperiment(
  assays = list(counts = count.mat),
  colData = sampC,
  rowTree = tinyTree
)

# add a minimal altExp: subset with 5 features
alt_counts <- count.mat[1:5, ]
alt_rowData <- S4Vectors::DataFrame(feature_type = rep("altFeature", 5))
alt_se <- SummarizedExperiment(assays = list(counts = alt_counts), rowData = alt_rowData)

altExp(tse, "altSubset") <- alt_se

 
colDataSpecs(tse)
}
