% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_enhance.R
\name{pMD_enhance}
\alias{pMD_enhance}
\title{Enhance parkinsonsMetagenomicData output}
\usage{
pMD_enhance(input.tse, data_type)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of
\code{loadMetagenomcData} or \code{returnSamples}}

\item{data_type}{\code{character}. `data_type` parameter used in 
\code{returnSamples}}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Formats and adds pieces to select outputs of parkinsonsMetagenomicData
}
\examples{

suppressMessages(library(TreeSummarizedExperiment))
data_types <- c("relative_abundance",
"pathabundance_unstratified", "genefamilies_unstratified")

# MetaPhlAn
data(Bedarf_pMD_raw_MetaPhlAn.tse)

tse_enhanced <- pMD_enhance(Bedarf_pMD_raw_MetaPhlAn.tse, 
  data_type = data_types[1])
  
# HUMAnN pathways
data(Bedarf_pMD_raw_HUMAnN_pwy.tse)

tse_enhanced <- pMD_enhance(Bedarf_pMD_raw_HUMAnN_pwy.tse, 
  data_type = data_types[2])
  
# HUMAnN Gene Families - example takes too long, see example in 
# `pMD_enhance_HUMAnN_genefam`
# data(Bedarf_pMD_raw_HUMAnN_GeneFam.tse)

# tse_enhanced <- pMD_enhance(Bedarf_pMD_raw_HUMAnN_GeneFam.tse, 
#   data_type = data_types[3])
}
