% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_enhance.R
\name{pMD_enhance_HUMAnN_genefam}
\alias{pMD_enhance_HUMAnN_genefam}
\title{Enhance HUMAnN unstratified gene families data}
\usage{
pMD_enhance_HUMAnN_genefam(input.tse)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData} or \code{returnSamples}. works exclusively for the
\code{data_type = "genefamilies_unstratified"} case}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance HUMAnN unstratified gene families data
}
\examples{
# this example takes a long time, but it's the problem of Gene Families:
# They are too many.

library(parkinsonsMetagenomicData)
data("sampleMetadata", package = "parkinsonsMetagenomicData")

data_types <- c("relative_abundance",
                "pathabundance_unstratified", "genefamilies_unstratified")

example_metadata.df <- sampleMetadata[grepl("Bedarf",
                                            sampleMetadata$study_name),]

tmp.tse <- returnSamples(example_metadata.df, data_type = data_types[3])

Bedarf_pMD_raw_HUMAnN_GeneFam.tse <- pMD_enhance_HUMAnN_genefam(tmp.tse)

Bedarf_pMD_raw_HUMAnN_GeneFam.tse
}
