% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{pMD_enhance_HUMAnN_genefam}
\alias{pMD_enhance_HUMAnN_genefam}
\title{Enhance HUMAnN unstratified gene families data}
\usage{
pMD_enhance_HUMAnN_genefam(input.tse, sampleMetadata.df)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData} or \code{returnSamples}. using 
`data_type = "genefamilies_unstratified"`}

\item{sampleMetadata.df}{\code{data.frame}. A data.frame to swap the default 
with. NB: it must have a `uuid` column with valid uuids}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance HUMAnN unstratified gene families data
}
\examples{
\dontrun{
library(parkinsonsMetagenomicData)

data_types <- c("relative_abundance", "pathabundance_unstratified", "genefamilies_unstratified")

hf_con <- accessParquetData(data_types = data_types)

example_metadata.df <- dplyr::filter(sampleMetadata, grepl("Bedarf", study_name))

# HUMAnN gene families
tse_basic <- loadParquetData(con = hf_con, data_type = "genefamilies_unstratified", filter_values = list(uuid = example_metadata.df$uuid))

tse_enhanced <- pMD_enhance(tse_basic, sampleMetadata.df = example_metadata.df, data_type = "genefamilies_unstratified", addPhyloTree = TRUE)

}
}
