% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{pMD_enhance_HUMAnN_pwy}
\alias{pMD_enhance_HUMAnN_pwy}
\title{Enhance HUMAnN unstratified pathways data}
\usage{
pMD_enhance_HUMAnN_pwy(input.tse, sampleMetadata.df)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData} or \code{returnSamples}. using 
`data_type = "pathabundance_unstratified"`}

\item{sampleMetadata.df}{\code{data.frame}. A data.frame to swap the default 
with. NB: it must have a `uuid` column with valid uuids}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance HUMAnN unstratified pathways data
}
\examples{

library(parkinsonsMetagenomicData)
data("sampleMetadata", package = "parkinsonsMetagenomicData")

data_types <- c("relative_abundance", 
"pathabundance_unstratified", "genefamilies_unstratified")

example_metadata.df <- sampleMetadata[grepl("Bedarf", 
sampleMetadata$study_name),]

tse_basic <- returnSamples(example_metadata.df, data_type = data_types[2])

tse_enhanced <- pMD_enhance(tse_basic, 
  sampleMetadata.df = example_metadata.df, 
  data_type = data_types[2])

}
