% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{pMD_enhance_MetaPhlAn}
\alias{pMD_enhance_MetaPhlAn}
\title{Enhance MetaPhlAn data}
\usage{
pMD_enhance_MetaPhlAn(input.tse, sampleMetadata.df)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of 
\code{loadMetagenomcData}
or \code{returnSamples} and `data_type = 'relative_abundance'`}

\item{sampleMetadata.df}{\code{data.frame}. A data.frame to swap the default 
with. NB: it must have a `uuid` column with valid uuids
to the \code{TreeSummarizedExperiment}? Works only with MetaPhlAn input 
(`data_type = "relative_abundance"`)}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance MetaPhlAn data
}
\examples{

library(parkinsonsMetagenomicData)
data("sampleMetadata", package = "parkinsonsMetagenomicData")
data_types <- c("relative_abundance", 
"pathabundance_unstratified", "genefamilies_unstratified")

example_metadata.df <- sampleMetadata[grepl("Bedarf", sampleMetadata$study_name),]

data_types <- c("relative_abundance", 
"pathabundance_unstratified", "genefamilies_unstratified")

tse_basic <- returnSamples(example_metadata.df, data_type = data_types[1])

tse_enhanced <- pMD_enhance(tse_basic, 
  sampleMetadata.df = example_metadata.df, 
  data_type = data_types[1])

}
