% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMD_data_wrangle.R
\name{pMD_enhance_MetaPhlAn}
\alias{pMD_enhance_MetaPhlAn}
\title{Enhance MetaPhlAn data}
\usage{
pMD_enhance_MetaPhlAn(
  input.tse,
  sampleMetadata.df,
  swap_metadata = FALSE,
  addPhyloTree = TRUE
)
}
\arguments{
\item{input.tse}{\code{TreeSummarizedExperiment}. as it comes out of \code{loadMetagenomcData}
or \code{returnSamples}}

\item{sampleMetadata.df}{\code{data.frame}. A data.frame to swap the default 
with. NB: it must have a `uuid` column with valid uuids}

\item{swap_metadata}{\code{logical}. Should the default metadata
be swapped for another. (default = `FALSE`)}

\item{addPhyloTree}{\code{logical}. Should a phylogenetic `rowTree` be added 
to the \code{TreeSummarizedExperiment}? Works only with MetaPhlAn input (`data_type = "relative_abundance"`)}

\item{data_type}{\code{character}. that got retrieved using 
\code{loadMetagenomcData} or \code{returnSamples}}
}
\value{
\code{TreeSummarizedExperiment}
}
\description{
Enhance MetaPhlAn data
}
\examples{
\dontrun{
library(parkinsonsMetagenomicData)

data_types <- c("relative_abundance", "pathabundance_unstratified", "genefamilies_unstratified")

hf_con <- accessParquetData(data_types = data_types)

example_metadata.df <- dplyr::filter(sampleMetadata, grepl("Bedarf", study_name))


# MetaPhlAn

tse_basic <- loadParquetData(con = hf_con, data_type = "relative_abundance", filter_values = list(uuid = example_metadata.df$uuid))

tse_enhanced <- pMD_enhance(tse_basic, sampleMetadata.df = example_metadata.df, data_type = "relative_abundance", swap_metadata = FALSE, addPhyloTree = TRUE)
}
}
