% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{read_TSE_from_dir}
\alias{read_TSE_from_dir}
\title{Read a TreeSummarizedExperiment from human readable files}
\usage{
read_TSE_from_dir(tse.dir)
}
\arguments{
\item{tse.dir}{\code{character}. The path that a `TreeSummarizedExperiment` 
was saved into}
}
\value{
A \code{TreeSummarizedExperiment}
}
\description{
This function is the reading counterpart of the `write_TSE_to_dir` function.
All files are human readable and maybe redundant, but at least more 
conservative.
}
\examples{
library(TreeSummarizedExperiment)
data("tinyTree", package = "TreeSummarizedExperiment")
set.seed(1234)

# basic assay
count.mat <- matrix(rpois(100, 50), nrow = 10)
rownames(count.mat) <- tinyTree$tip.label
colnames(count.mat) <- paste0("C_", 1:10)

# sample data
sampC <- data.frame(
  condition = factor(rep(c("control", "trt"), each = 5), 
                     levels = c("trt", "control")),
  gender = factor(sample(c("male", "female"), size = 10, replace = TRUE),
                  levels = c("male", "female"))
)
rownames(sampC) <- colnames(count.mat)

# create rowData
rowData <- DataFrame(
Feature_gr1 = factor(rep(c("blue", "red"), each = 5)),
Feature_gr2 = rep(c(2,3,4,6,1), 2)
)
rownames(rowData) <- rownames(count.mat)

# base tse
tse <- TreeSummarizedExperiment(
  assays = list(counts = count.mat),
  colData = sampC,
  rowData = rowData,
  rowTree = tinyTree
)

# add altExp
alt_colData <- sampC
alt_colData$altTreat <- rep(0:1, 5)

alt_counts <- count.mat[1:5, ]
alt_rowData <- S4Vectors::DataFrame(feature_type = rep("altFeature", 5))
alt_se <- TreeSummarizedExperiment(assays = list(counts = alt_counts), rowData = alt_rowData, colData = alt_colData)

altExp(tse, "altSubset") <- alt_se
metadata(tse)$info <- "test_metadata"

tmpdir <- file.path(tempdir(), "testTSE_complete2")
unlink(tmpdir, recursive = TRUE)
write_TSE_to_dir(tse, tmpdir)
list.files(tmpdir, recursive = TRUE)

# re-load tse
tse2 <- read_TSE_from_dir(tmpdir)
all.equal(tse, tse2)  # TRUE
}
