% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{read_TSE_from_dir_noAltExp}
\alias{read_TSE_from_dir_noAltExp}
\title{Read multiple tsv files into a TreeSummarizedExperiment}
\usage{
read_TSE_from_dir_noAltExp(tse.dir)
}
\arguments{
\item{tse.dir}{\code{character} indicating the directory that was created
with `write_TSE_to_dir()`}
}
\value{
A \code{TreeSummarizedExperiment}
}
\description{
This utility function is the counterpart of
}
\examples{
library(TreeSummarizedExperiment)
data("tinyTree", package = "TreeSummarizedExperiment")
set.seed(1234)

# basic assay
count.mat <- matrix(rpois(100, 50), nrow = 10)
rownames(count.mat) <- tinyTree$tip.label
colnames(count.mat) <- paste0("C_", 1:10)

# sample data
sampC <- data.frame(
  condition = factor(rep(c("control", "trt"), each = 5), 
                     levels = c("trt", "control")),
  gender = factor(sample(c("male", "female"), size = 10, replace = TRUE),
                  levels = c("male", "female"))
)
rownames(sampC) <- colnames(count.mat)

# create rowData
rowData <- DataFrame(
Feature_gr1 = factor(rep(c("blue", "red"), each = 5)),
Feature_gr2 = rep(c(2,3,4,6,1), 2)
)
rownames(rowData) <- rownames(count.mat)

# base tse
tse <- TreeSummarizedExperiment(
  assays = list(counts = count.mat),
  colData = sampC,
  rowData = rowData,
  rowTree = tinyTree
)

# add altExp
alt_colData <- sampC
alt_colData$altTreat <- rep(0:1, 5)

alt_counts <- count.mat[1:5, ]
alt_rowData <- S4Vectors::DataFrame(feature_type = rep("altFeature", 5))
alt_se <- TreeSummarizedExperiment(assays = list(counts = alt_counts), rowData = alt_rowData, colData = alt_colData)

altExp(tse, "altSubset") <- alt_se
metadata(tse)$info <- "test_metadata"

tmpdir <- file.path(tempdir(), "testTSE3")
dir.create(tmpdir)
write_TSE_to_dir_noAltExp(tse, tmpdir)
list.files(tmpdir, recursive = TRUE)

# re-load tse
tse2 <- read_TSE_from_dir_noAltExp(tmpdir)
tse2
tse3 <- tse
altExp(tse3) <- NULL
all.equal(tse3, tse2)  # TRUE
}
