% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{read_TSE_from_dir_noAltExp}
\alias{read_TSE_from_dir_noAltExp}
\title{Read multiple tsv files into a TreeSummarizedExperiment}
\usage{
read_TSE_from_dir_noAltExp(tse.dir)
}
\arguments{
\item{tse.dir}{\code{character} indicating the directory that was created
with `write_TSE_to_dir()`}
}
\value{
A \code{TreeSummarizedExperiment}
}
\description{
This utility function is the counterpart of
}
\examples{
 library(TreeSummarizedExperiment)
 data("tinyTree")

 # the count.mat table
 count.mat <- matrix(rpois(100, 50), nrow = 10)

 rownames(count.mat) <- c(tinyTree$tip.label)

 colnames(count.mat) <- paste("C_", 1:10, sep = "_")

 # The sample information
sampC <- data.frame(
  condition = rep(c("control", "trt"), each = 5),
  gender = as.numeric(sample(x = 1:2, size = 10, replace = TRUE))
  )

rownames(sampC) <- colnames(count.mat)

 # build a TreeSummarizedExperiment object
 tse <- TreeSummarizedExperiment(
   assays = list("counts" = count.mat),
   colData = sampC,
   rowTree = tinyTree
   )
 tse <- reorder_taxa_with_phyloTree_labels(tse)
 # save the tse

 set.seed(1234)

 write_TSE_to_dir(tse, tempdir())

 list.files(file.path(tempdir(), "tse"), recursive = TRUE)

 tse2 <- read_TSE_from_dir(file.path(tempdir(), "tse"))

 identical(tse, tse2) # FALSE
 all.equal(tse, tse2) # TRUE
 identical(rownames(tse), rownames(tse2)) # TRUE
 identical(colnames(tse), colnames(tse2)) # TRUE
 identical(tse@rowTree$phylo$tip.label, tse2@rowTree$phylo$tip.label) # TRUE
 identical(tse@rowTree$phylo$node.label, tse2@rowTree$phylo$node.label) # TRUE
 identical(
 round(tse@rowTree$phylo$edge.length,10),
 round(tse2@rowTree$phylo$edge.length, 10)
 ) # TRUE, identical at least to the 10th digit

 identical(assay(tse), assay(tse)) # FALSE, I don't yet know why
 identical(round(assay(tse), 10), round(assay(tse), 10)) # The encoding
 # could be improved, but the matrices are practically the same
}
