% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_metaphlan.R
\name{wrangle_metaphlan}
\alias{wrangle_metaphlan}
\title{wrangle a metaphlan profiles table}
\usage{
wrangle_metaphlan(mpa, taxonomic_lvl = "Species")
}
\arguments{
\item{mpa}{a Metaphlan matrix as a data frame}

\item{taxonomic_lvl}{\code{character} indicating the taxonomic level wanted. Default is `Species`}
}
\value{
a \code{list containing both `profiles` and `taxonomies`}
}
\description{
wrangle a metaphlan profiles table
}
\examples{

data("metaphlanTestData")
wrangled_metaphlans <- sapply(c("Kingdom", "Class","Species"),
function(taxLvl)
wrangle_metaphlan(metaphlanTestData, taxonomic_lvl = taxLvl), 
USE.NAMES = TRUE, simplify = FALSE)
replace original names with those that were gradually added going up the
taxonomic tree
}
