% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_metaphlan.R
\name{wrangle_metaphlan}
\alias{wrangle_metaphlan}
\title{wrangle a metaphlan profiles table}
\usage{
wrangle_metaphlan(mpa, taxonomic_lvl = "Species")
}
\arguments{
\item{mpa}{a Metaphlan matrix as a data frame as comes out after 
`merge_metaphlan.py`}

\item{taxonomic_lvl}{\code{character} indicating the taxonomic level wanted. 
Default is `Species`}
}
\value{
a \code{list containing both `profiles` and `taxonomies`}
}
\description{
Works really similarly to mia::importMetaphlan, the difference being that 
this function starts from data already loaded in R. Also, it does not 
handel ncbi_tax_id so well
}
\examples{

wallen_profiles <- data.table::fread(system.file("extdata",
                                                 "WallenZD_2022_metaphlan3_profiles.tsv.bz2",
                                                 package = "biobakeryUtils"))


profiles <- sapply(c("Phylum", "Genus", "Species"), function(taxLvl) {
  wrangle_metaphlan(wallen_profiles, taxonomic_lvl = taxLvl)
},
USE.NAMES = TRUE, simplify = FALSE)

profiles$Species$profiles[1:5,1:5]
profiles$Species$taxonomies[1:5,]

}
