% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{write_TSE_to_dir}
\alias{write_TSE_to_dir}
\title{Split a `TreeSummarizedExperiment` object into multiple human readable files}
\usage{
write_TSE_to_dir(tse, out.dir)
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} object with or without
rowTree and/or colTree}

\item{out.dir}{\code{character} specifying the directory in which a new directory
will be generated, with the name of the R object passed to the `tse`
parameter.}
}
\value{
Nothing, the main output of this function is multiple .tsv files
}
\description{
The components are saved into a directory that gets created by this function.
AltExp slot is ignored in this function
}
\examples{
library(TreeSummarizedExperiment)
data("tinyTree", package = "TreeSummarizedExperiment")
set.seed(1234)

# basic assay
count.mat <- matrix(rpois(100, 50), nrow = 10)
rownames(count.mat) <- tinyTree$tip.label
colnames(count.mat) <- paste0("C_", 1:10)

# sample data
sampC <- data.frame(
  condition = factor(rep(c("control", "trt"), each = 5), 
                     levels = c("trt", "control")),
  gender = factor(sample(c("male", "female"), size = 10, replace = TRUE),
                  levels = c("male", "female"))
)
rownames(sampC) <- colnames(count.mat)

# create rowData
rowData <- DataFrame(
Feature_gr1 = factor(rep(c("blue", "red"), each = 5)),
Feature_gr2 = rep(c(2,3,4,6,1), 2)
)
rownames(rowData) <- rownames(count.mat)

# base tse
tse <- TreeSummarizedExperiment(
  assays = list(counts = count.mat),
  colData = sampC,
  rowData = rowData,
  rowTree = tinyTree
)

# add altExp
alt_colData <- sampC
alt_colData$altTreat <- rep(0:1, 5)

alt_counts <- count.mat[1:5, ]
alt_rowData <- S4Vectors::DataFrame(feature_type = rep("altFeature", 5))
alt_se <- TreeSummarizedExperiment(assays = list(counts = alt_counts), rowData = alt_rowData, colData = alt_colData)

altExp(tse, "altSubset") <- alt_se
metadata(tse)$info <- "test_metadata"

tmpdir <- file.path(tempdir(), "testTSE_complete")
write_TSE_to_dir(tse, tmpdir)
list.files(tmpdir, recursive = TRUE)
}
