% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperimentUtils.R
\name{write_TSE_to_dir_noAltExp}
\alias{write_TSE_to_dir_noAltExp}
\title{Split a `TreeSummarizedExperiment` object into multiple human readable files}
\usage{
write_TSE_to_dir_noAltExp(tse, out.dir)
}
\arguments{
\item{tse}{\code{TreeSummarizedExperiment} object with or without
rowTree and/or colTree}

\item{out.dir}{\code{character} specifying the directory in which a new directory
will be generated, with the name of the R object passed to the `tse`
parameter.}
}
\value{
Nothing, the main output of this function is multiple .tsv files
}
\description{
The components are saved into a directory that gets created by this function.
AltExp slot is ignored in this function
}
\examples{
 library(TreeSummarizedExperiment)
 data("tinyTree")

 # the count.mat table
 count.mat <- matrix(rpois(100, 50), nrow = 10)

 rownames(count.mat) <- c(tinyTree$tip.label)

 colnames(count.mat) <- paste("C_", 1:10, sep = "_")

 # The sample information
sampC <- data.frame(
  condition = rep(c("control", "trt"), each = 5),
  gender = sample(x = 1:2, size = 10, replace = TRUE)
  )

rownames(sampC) <- colnames(count.mat)

 # build a TreeSummarizedExperiment object
 tse <- TreeSummarizedExperiment(
   assays = list("counts" = count.mat),
   colData = sampC,
   rowTree = tinyTree
   )

 # write the tse

 set.seed(1234)

 write_TSE_to_dir_noAltExp(tse, tempdir())

 list.files(file.path(tempdir(), "tse"), recursive = TRUE)
}
